Server/app/requirements.txt
fastapi
uvicorn
pandas
python-docx
numpy
passlib[bcrypt]
python-multipart
mysql-connector-python
dotenv
PyJWT